creative_analytics/creative_analytics_agents/requirements.txt
fastapi==0.118.3
uvicorn==0.38.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
pandas==2.3.3
numpy==2.3.5
google-adk==1.19.0
google-genai==1.52.0
google-cloud-bigquery==3.38.0
google-cloud-storage==3.6.0
httpx==0.28.1
tenacity==9.1.2